Add s3 key prefix for DAGs

+++ src/python/beeflow/experiments/internal/services/bucket_manager/bucket_manager.py
@@ -2,6 +2,7 @@ import abc
 import os
 from os import listdir
 from os.path import isfile, join
+from typing import Optional
 
 from mypy_boto3_s3.service_resource import Bucket
 
@@ -17,11 +18,15 @@ class IBucketManager(abc.ABC):
 
 
 class BucketManager(IBucketManager):
-    def __init__(self, bucket: Bucket):
+    def __init__(self, bucket: Bucket, s3_dags_path: Optional[str]):
         self.bucket = bucket
+        self.s3_dags_path = s3_dags_path
 
     def clear_dags(self) -> None:
-        self.bucket.objects.all().delete()
+        if self.s3_dags_path is None:
+            self.bucket.objects.all().delete()
+            return
+        self.bucket.objects.filter(Prefix=f"{self.s3_dags_path}").delete()
 
     def publish_dags(self, dags_path: str) -> [str]:
         if not os.path.isdir(dags_path):
@@ -29,9 +34,10 @@ class BucketManager(IBucketManager):
 
         files = [f for f in listdir(dags_path) if isfile(join(dags_path, f)) and f not in ["BUILD"]]
         for file in files:
+            s3_key = f"{self.s3_dags_path}/{file}" if self.s3_dags_path is not None else file
             self.bucket.upload_file(
                 Filename=os.path.join(dags_path, file),
-                Key=file,
+                Key=s3_key,
             )
 
         return files
